Dissolve remaining style module into css.py
